uploaded initial parser
Readme file description updated.
First version of the parser uploaded.

The parser currently takes a file as a parameter and breaks it up into
its component pieces. The source code for the 0.12-a trunk version was
used to develop the parser. The parser is currently incomplete as some
of the sections are being ignored.

Of particular note, the Screenshot set is unable to be identified, and a
few of the later segments have not yet had their lineNum end section
coded.