First cut at exporting to HTML
Creating my own code to export and write to an html file.

forensicsHTML created as a test code base. This will be completely
rewritten as I get familiar with exporting the output from
gameCollection into a html file.

style.css blatently stolen off the crawl scoring website and will be
modified as needed.
test.html will be a work in progress. I will attempt to not save it for
the sake of saving space, updating only every once in a while.