end information extracted
Hopefully this encapsulates all possible end results. The largest
optimization will be checking to see where the death piece stops.

One brainstorm possibility is to keep looping until it finds "the game
lasted" and then backtrack the location information (Is it only one
line?) and then use the rest as the end message.