revised achievement structure
Achievements are now set up in the following way:
A dictionary entry w/ ID, short name, long name, short desc, long desc.
The achievementList in gameCollection is now a list of lists, with each
element's [0] being True or False, indicating whether or not the
achievement is complete and the rest being comments / progress
information.